cleaned up npm
