+++ ext/npm-extension.js
@@ -1,91 +1,9 @@
 "format cjs";
 
+var utils = require("./npm-utils");
+exports.includeInBuild = true;
 
-// HELPERS USED BY npm.js
-function isRelative(path) {
-	return  path.substr(0,1) === ".";
-}
-function pkgMain(pkg) {
-	return  removeJS( (pkg.system && pkg.system.main) || (typeof pkg.browser === "string" && pkg.browser) || pkg.main || 'index' ) ;
-}
-function removeJS(path) {
-	return path.replace(/\.js(!|$)/,function(whole, part){return part;});
-}
-
-// Converts a parsed module name to a string
-function createModuleName (descriptor, standard) {
-	if(standard) {
-		return descriptor.moduleName;
-	} else {
-		var modulePath;
-		if(descriptor.modulePath) {
-			modulePath = descriptor.modulePath.substr(0,2) === "./" ? descriptor.modulePath.substr(2) : descriptor.modulePath;
-		}
-		return descriptor.packageName
-			+ (descriptor.version ? '@' + descriptor.version : '')
-			+ (modulePath ? '#' + modulePath : '')
-			+ (descriptor.plugin ? descriptor.plugin : '');
-	}
-};
-
-// Breaks a string moduleName into parts.
-// packageName@version!plugin#modulePath
-// "./lib/bfs"
-function parseModuleName (moduleName, currentPackageName) {
-	var pluginParts = moduleName.split('!');
-	var modulePathParts = pluginParts[0].split("#");
-	var versionParts = modulePathParts[0].split("@");
-	// it could be something like `@empty`
-	if(!modulePathParts[1] && !versionParts[0]) {
-		versionParts = ["@"+versionParts[0]];
-	}
-	var packageName, 
-		modulePath;
-	
-	// if relative, use currentPackageName
-	if( currentPackageName && isRelative(moduleName) ) {
-		packageName= currentPackageName;
-		modulePath = versionParts[0];
-	} else {
-		
-		if(modulePathParts[1]) { // foo@1.2#./path
-			packageName = versionParts[0];
-			modulePath = modulePathParts[1];
-		} else {
-			// test/abc
-			var folderParts = versionParts[0].split("/");
-			packageName = folderParts.shift();
-			modulePath = folderParts.join("/");
-		}
-		
-	}
-	
-	return {
-		plugin: pluginParts.length === 2 ? "!"+pluginParts[1] : undefined,
-		version: versionParts[1],
-		modulePath: modulePath,
-		packageName: packageName,
-		moduleName: moduleName
-	};
-}
-
-function childPackageAddress(parentPackageAddress, childName){
-	var packageFolderName = parentPackageAddress.replace(/\/package\.json.*/,"");
-	return (packageFolderName ? packageFolderName+"/" : "")+"node_modules/" + childName + "/package.json";
-}
-
-
-function parentNodeModuleAddress(address) {
-	var nodeModules = "/node_modules/",
-		nodeModulesIndex = address.lastIndexOf(nodeModules),
-		prevModulesIndex = address.lastIndexOf(nodeModules, nodeModulesIndex-1);
-	if(prevModulesIndex >= 0) {
-		return address.substr(0, prevModulesIndex+nodeModules.length - 1 );
-	}
-}
-
-var extension = function(System){
-	
+exports.addExtension = function(System){
 	/**
 	 * Normalize has to deal with a "tricky" situation.  There are module names like
 	 * "css" -> "css" normalize like normal
@@ -101,27 +19,30 @@ var extension = function(System){
 	var oldNormalize = System.normalize;
 	System.normalize = function(name, parentName, parentAddress){
 		
-		var refPkg = findPackageByAddress(this, parentName, parentAddress);
+		// Get the current package
+		var refPkg = utils.pkg.findByModuleNameOrAddress(this, parentName, parentAddress);
 		
 		// this isn't in a package, so ignore
 		if(!refPkg) {
 			return oldNormalize.call(this, name, parentName, parentAddress);
 		}
-		// TODO: joining ...
-		var parsedModuleName = parsedModuleNameFromPackage(this, refPkg, name, parentName);
 		
-		var depPkg = findDepPackage(this, refPkg, parsedModuleName.packageName);
+		// Using the current package, get info about what it is probably asking for
+		var parsedModuleName = utils.moduleName.parseFromPackage(this, refPkg, name, parentName);
+		
+		// Look for the dependency package specified by the current package
+		var depPkg = utils.pkg.findDep(this, refPkg, parsedModuleName.packageName);
 		
 		// This really shouldn't happen, but lets find a package.
 		if (!depPkg) {
-			depPkg = findPackage(this, parsedModuleName.packageName);
+			depPkg = utils.pkg.findByName(this, parsedModuleName.packageName);
 		}
 		// It could be something like `fs` so check in globals
 		if(!depPkg) {
 			var browserPackageName = this.globalBrowser[parsedModuleName.packageName];
 			if(browserPackageName) {
 				parsedModuleName.packageName = browserPackageName;
-				depPkg = findPackage(this, parsedModuleName.packageName);
+				depPkg = utils.pkg.findByName(this, parsedModuleName.packageName);
 			}
 		}
 		
@@ -129,32 +50,27 @@ var extension = function(System){
 			parsedModuleName.version = depPkg.version;
 			// add the main path
 			if(!parsedModuleName.modulePath) {
-				parsedModuleName.modulePath = pkgMain(depPkg);
+				parsedModuleName.modulePath = utils.pkg.main(depPkg);
 			}
-			return oldNormalize.call(this, createModuleName(parsedModuleName), parentName, parentAddress);
+			return oldNormalize.call(this, utils.moduleName.create(parsedModuleName), parentName, parentAddress);
 		} else {
 			if(depPkg === this.npmPaths.__default) {
-				//debugger;
-				var localName = parsedModuleName.modulePath ? parsedModuleName.modulePath+(parsedModuleName.plugin? parsedModuleName.plugin: "") : pkgMain(depPkg);
+				// if the current package, we can't? have the
+				// module name look like foo@bar#./zed
+				var localName = parsedModuleName.modulePath ? 
+					parsedModuleName.modulePath+(parsedModuleName.plugin? parsedModuleName.plugin: "") : 
+					utils.pkg.main(depPkg);
 				return oldNormalize.call(this, localName, parentName, parentAddress);
 			}
 			return oldNormalize.call(this, name, parentName, parentAddress);
 		}
 		
 	};
-	function makeRelative(path){
-		if( isRelative(path) && path.substr(0,1) !== "/" ) {
-			return path;
-		} else {
-			return "./"+path;
-		}
-	}
-
 	
 	
 	var oldLocate = System.locate;
 	System.locate = function(load){
-		var parsedModuleName = parseModuleName(load.name),
+		var parsedModuleName = utils.moduleName.parse(load.name),
 			loader = this;
 		
 		// @ is not the first character
@@ -162,17 +78,12 @@ var extension = function(System){
 			var pkg = this.npm[parsedModuleName.packageName];
 			if(pkg) {
 				return oldLocate.call(this, load).then(function(address){
-					var root = pkg === loader.npmPaths.__default ?
-						removePackage( pkg.fileUrl ) :
-						packageFolderAddress(pkg.fileUrl);
 					
-					var lib = pkg.system && pkg.system.directories && pkg.system.directories.lib;
-					if(lib) {
-						root = joinURL(root, lib);
-					}
+					var root = utils.pkg.rootDir(pkg, pkg === loader.npmPaths.__default);
+					
 					
 					if(parsedModuleName.modulePath) {
-						return joinURL( root, addJS(parsedModuleName.modulePath));
+						return utils.path.joinURIs( utils.path.addEndingSlash(root), utils.path.addJS(parsedModuleName.modulePath));
 					} 
 					
 					return address;
@@ -182,187 +93,4 @@ var extension = function(System){
 		return oldLocate.call(this, load);
 	};
 
-	var extend = function(d, s){
-		for(var prop in s) {
-			d[prop] = s[prop];
-		}
-		return d;
-	};
-
-	// PACKAGE HELPERS ========================
-	// Help locate modules based on packages info
-	function findPackageByAddress(loader, parentName, parentAddress) {
-		if(loader.npm) {
-			if(parentName) {
-				var parsed = parseModuleName(parentName);
-				if(parsed.version && parsed.packageName) {
-					var name = parsed.packageName+"@"+parsed.version;
-					if(name in loader.npm) {
-						return loader.npm[name];
-					}
-				}
-			}
-			if(parentAddress) {
-				var packageFolder = packageFolderAddress(parentAddress);
-				return packageFolder ? loader.npmPaths[packageFolder] : loader.npmPaths.__default;
-			} else {
-				return loader.npmPaths.__default;
-			}
-		}
-		
-	}
-	function findPackage(loader, name) {
-		if(loader.npm && !startsWithDotSlash(name)) {
-			return loader.npm[name];
-		}
-	}
-	function findDepPackage(loader, refPackage, name) {
-		if(loader.npm && refPackage && !startsWithDotSlash(name)) {
-			// Todo .. first part of name
-			var curPackage = childPackageAddress(refPackage.fileUrl, name).replace(/\/package\.json.*/,"");
-			while(curPackage) {
-				var pkg = loader.npmPaths[curPackage];
-				if(pkg) {
-					return pkg;
-				}
-				var parentAddress = parentNodeModuleAddress(curPackage);
-				if(!parentAddress) {
-					return;
-				}
-				curPackage = parentAddress+"/"+name;
-			}
-		}
-	}
-	
-	function parsedModuleNameFromPackage(loader, refPkg, name, parentName) {
-		var packageName = (refPkg.system && refPkg.system.name) || refPkg.name,
-		    parsedModuleName = parseModuleName(name, packageName);
-		    
-		if( isRelative( parsedModuleName.modulePath ) ) {
-			var parentParsed = parseModuleName( parentName, packageName );
-			if( parentParsed.packageName === parsedModuleName.packageName && parentParsed.modulePath ) {
-				parsedModuleName.modulePath = makeRelative( joinURIs(parentParsed.modulePath, parsedModuleName.modulePath) );
-			}
-		}
-		var mapName = createModuleName(parsedModuleName),
-		    mappedName;
-		
-		// The refPkg might have a browser [https://github.com/substack/node-browserify#browser-field] mapping.
-		// Perform that mapping here.
-		if(refPkg.browser && (mapName in refPkg.browser)  && (!refPkg.system || !refPkg.system.ignoreBrowser)) {
-			mappedName = refPkg.browser[mapName] === false ? "@empty" : refPkg.browser[mapName];
-		}
-		// globalBrowser looks like: {moduleName: aliasName, pgk: aliasingPkg}
-		var global = loader && loader.globalBrowser && loader.globalBrowser[mapName];
-		if(global) {
-			mappedName = global.moduleName === false ? "@empty" : global.moduleName;
-		}
-		
-		if(mappedName) {
-			return parseModuleName(mappedName, packageName);
-		} else {
-			return parsedModuleName;
-		}
-	}
-
-	function addJS(path){
-		if(/\.\w+$/.test(path)) {
-			return path;
-		} else {
-			return path+".js";
-		}
-	}
-
-	function removePackage(path){
-		return path.replace(/\/package\.json.*/,"");
-	}
-	
-	function packageFolderAddress(address){
-		var nodeModules = "/node_modules/",
-			nodeModulesIndex = address.lastIndexOf(nodeModules),
-			nextSlash = address.indexOf("/", nodeModulesIndex+nodeModules.length);
-		if(nodeModulesIndex >= 0) {
-			return nextSlash>=0 ? address.substr(0, nextSlash) : address;
-		}
-	}
-	function parseURI(url) {
-		var m = String(url).replace(/^\s+|\s+$/g, '').match(/^([^:\/?#]+:)?(\/\/(?:[^:@]*(?::[^:@]*)?@)?(([^:\/?#]*)(?::(\d*))?))?([^?#]*)(\?[^#]*)?(#[\s\S]*)?/);
-			// authority = '//' + user + ':' + pass '@' + hostname + ':' port
-			return (m ? {
-			href     : m[0] || '',
-			protocol : m[1] || '',
-			authority: m[2] || '',
-			host     : m[3] || '',
-			hostname : m[4] || '',
-			port     : m[5] || '',
-			pathname : m[6] || '',
-			search   : m[7] || '',
-			hash     : m[8] || ''
-		} : null);
-	}
-	// TODO: merge with joinURL
-	function joinURIs(base, href) {
-		function removeDotSegments(input) {
-			var output = [];
-			input.replace(/^(\.\.?(\/|$))+/, '')
-				.replace(/\/(\.(\/|$))+/g, '/')
-				.replace(/\/\.\.$/, '/../')
-				.replace(/\/?[^\/]*/g, function (p) {
-					if (p === '/..') {
-						output.pop();
-					} else {
-						output.push(p);
-					}
-				});
-			return output.join('').replace(/^\//, input.charAt(0) === '/' ? '/' : '');
-		}
-	
-		href = parseURI(href || '');
-		base = parseURI(base || '');
-	
-		return !href || !base ? null : (href.protocol || base.protocol) +
-			(href.protocol || href.authority ? href.authority : base.authority) +
-			removeDotSegments(href.protocol || href.authority || href.pathname.charAt(0) === '/' ? href.pathname : (href.pathname ? ((base.authority && !base.pathname ? '/' : '') + base.pathname.slice(0, base.pathname.lastIndexOf('/') + 1) + href.pathname) : base.pathname)) +
-				(href.protocol || href.authority || href.pathname ? href.search : (href.search || base.search)) +
-				href.hash;
-	}
-	function startsWithDotSlash( path ) {
-		return path.substr(0,2) === "./";
-	}
-	function endsWithSlash(path){
-		return path[path.length -1] === "/";
-	}
-	function removeTrailingSlash( path ) {
-		if(endsWithSlash(path)) {
-			return path.substr(0, path.length -1);
-		} else {
-			return path;
-		}
-	}
-	function removeLeadingDotSlash( path ) {
-		if(startsWithDotSlash(path)) {
-			return path.substr(2);
-		} else {
-			return path;
-		}
-	}
-	// TODO: merge joinURIs
-	function joinURL(baseURL, url){
-		baseURL = removeTrailingSlash(baseURL);
-		url = removeLeadingDotSlash(url);
-		return baseURL+"/"+url;
-	}
-};
-
-
-exports.createModuleName = createModuleName;
-exports.parseModuleName = parseModuleName;
-exports.childPackageAddress = childPackageAddress;
-exports.parentNodeModuleAddress= parentNodeModuleAddress;
-exports.isRelative= isRelative;
-exports.pkgMain = pkgMain;
-
-exports.out = function(){
-	return [createModuleName,parseModuleName,isRelative, removeJS, pkgMain, childPackageAddress, parentNodeModuleAddress].join("\n")+"\n"+
-		"("+extension.toString()+")(loader);\n";
 };
