+++ ext/npm.js
@@ -1,27 +1,15 @@
 "format cjs";
 
 // TODO: cleanup removing package.json
-var SemVer = require('./semver');
-var npmExtension = require('./npm-extension');
-var createModuleName = npmExtension.createModuleName;
-var parseModuleName = npmExtension.parseModuleName;
+var utils = require('./npm-utils');
+var crawl = require('./npm-crawl');
+
 
 // Add @loader, for SystemJS
 if(!System.has("@loader")) {
 	System.set('@loader', System.newModule({'default':System, __useDefault: true}));
 }
 
-
-// BAISC HELPERS
-var extend = function(d, s){
-	for(var prop in s) {
-		d[prop] = s[prop];
-	}
-	return d;
-};
-function truthy(x) {
-	return x;
-}
 
 
 
@@ -50,9 +38,9 @@ exports.translate = function(load){
 	};
 	var pkg = {origFileUrl: load.address, fileUrl: load.address};
 	
-	processPkg(context, pkg, load.source);
+	crawl.processPkgSource(context, pkg, load.source);
 	
-	return processDeps(context, pkg).then(function(){
+	return crawl.deps(context, pkg, true).then(function(){
 		// clean up packages so everything is unique
 		var names = {};
 		var packages = [];
@@ -73,184 +61,13 @@ exports.translate = function(load){
 				packages[pkg.name+"@"+pkg.version] = true;
 			}
 		});
-		return "define(['@loader'], function(loader){\n" +
-			npmExtension.out()+
-		    (pkg.main ? "if(!System.main){ System.main = "+JSON.stringify(npmExtension.pkgMain(pkg))+"; }\n" : "") + 
+		return "define(['@loader','npm-extension'], function(loader, npmExtension){\n" +
+			"npmExtension.addExtension(loader);\n"+
+		    (pkg.main ? "if(!System.main){ System.main = "+JSON.stringify(utils.pkg.main(pkg))+"; }\n" : "") + 
 			"("+translateConfig.toString()+")(loader, "+JSON.stringify(packages, null, " ")+");\n" +
 		"});";
 	});
 };
-
-
-// CRAWLING HELPERS =========================
-// Helpers that read through package.json
-
-// processes a package.json text, but not its dependencies.
-function processPkg(context, pkg, source) {
-	var packageJSON = JSON.parse(source);
-	extend(pkg, packageJSON);
-	context.packages.push(pkg);
-	return pkg;
-}
-
-function nodeModuleAddress(address) {
-	if(!address) debugger;
-	var nodeModules = "/node_modules/",
-		nodeModulesIndex = address.lastIndexOf(nodeModules);
-	if(nodeModulesIndex >= 0) {
-		return address.substr(0, nodeModulesIndex+nodeModules.length - 1 );
-	}
-}
-// processes a package.json's dependencies
-function processDeps(context, pkg) {
-
-	var deps = getDependencies(context.loader, pkg);
-	return Promise.all(deps.map(function(childPkg){
-		if(childPkg._isPeerDependency && nodeModuleAddress(pkg.fileUrl) ) {
-			childPkg.origFileUrl = nodeModuleAddress(pkg.fileUrl)+"/"+childPkg.name+"/package.json";
-		} else {
-			childPkg.origFileUrl = npmExtension.childPackageAddress(pkg.fileUrl, childPkg.name);
-		}
-		
-		
-		// check if childPkg matches a parent's version ... if it does ... do nothing
-		if(hasParentPackageThatMatches(context, childPkg)) {
-			return;
-		}
-		
-		if(isSameRequestedVersionFound(context, childPkg)) {
-			return;
-		}
-		
-		
-		
-		// otherwise go get child ... but don't process dependencies until all of these dependencies have finished
-		return npmLoad(context, childPkg).then(function(source){
-			if(source) {
-				return processPkg(context, childPkg, source);
-			} // else if there's no source, it's likely because this dependency has been found elsewhere
-		});
-		
-	}).filter(truthy)).then(function(packages){
-		// at this point all dependencies of pkg have been loaded, it's ok to get their children
-
-		return Promise.all(packages.map(function(childPkg){
-			if(childPkg) {
-				return processDeps(context, childPkg);
-			} 
-		}).filter(truthy));
-	});
-}
-
-
-function isSameRequestedVersionFound(context, childPkg) {
-	if(!context.versions[childPkg.name]) {
-		context.versions[childPkg.name] = {};
-	}
-	var versions = context.versions[childPkg.name];
-	
-	var requestedRange = childPkg.version;
-	
-	if( !SemVer.validRange(childPkg.version) ) {
-		
-		if(/^[\w_\-]+\/[\w_\-]+(#[\w_\-]+)?$/.test(requestedRange)  ) {
-			
-			requestedRange = "git+https://github.com/"+requestedRange;
-			if(!/(#[\w_\-]+)?$/.test(requestedRange)) {
-				requestedRange += "#master";
-			}
-		}
-	}
-	var version = versions[requestedRange];
-	
-	if(!version) {
-		versions[requestedRange] = childPkg;
-	} else {
-		// add a placeholder at this path
-		context.paths[childPkg.origFileUrl] = version;
-		return true;
-	}
-}
-
-function hasParentPackageThatMatches(context, childPkg){
-	// check paths
-	var parentAddress = npmExtension.parentNodeModuleAddress(childPkg.origFileUrl);
-	while(parentAddress) {
-		var packageAddress = parentAddress+"/"+childPkg.name+"/package.json";
-		var parentPkg = context.paths[packageAddress];
-		if(parentPkg) {
-			if(SemVer.satisfies(parentPkg.version, childPkg.version)) {
-				return parentPkg;
-			}
-		}
-		parentAddress = npmExtension.parentNodeModuleAddress(packageAddress);
-	}
-}
-
-function addDeps(packageJSON, dependencies, deps, defaultProps){
-	for(var name in dependencies) {
-		if(!packageJSON.system || !packageJSON.system.npmIgnore || !packageJSON.system.npmIgnore[name]) {
-			deps[name] = extend(defaultProps || {}, {name: name, version: dependencies[name]});
-		}
-	}
-}
-function getDependencyMap(loader, packageJSON, isRoot){
-	var deps = {};
-	
-	addDeps(packageJSON, packageJSON.peerDependencies || {}, deps, {_isPeerDependency: true});
-	addDeps(packageJSON, packageJSON.dependencies || {}, deps);
-	// Only get the devDependencies if this is the root bower and the 
-	// `npmDev` option is enabled
-	if(loader.npmDev && (isRoot || !loader._npmMainLoaded)) {
-		addDeps(packageJSON, packageJSON.devDependencies || {}, deps);
-		loader._npmMainLoaded = true;
-	}
-	return deps;
-}
-// Combines together dependencies and devDependencies (if npmDev option is enabled)
-function getDependencies(loader, packageJSON){
-	var deps = getDependencyMap(loader, packageJSON);
-	
-	var dependencies = [];
-	for(var name in deps) {
-		dependencies.push(deps[name]);
-	}
-	
-	return dependencies;
-};
-
-// Loads package.json
-// if it finds one, it sets that package in paths
-// so it won't be loaded twice.
-function npmLoad(context, pkg, fileUrl){
-	fileUrl = fileUrl || pkg.origFileUrl;
-	return System.fetch({
-		address: fileUrl,
-		name: fileUrl,
-		metadata: {}
-	}).then(function(source){
-		context.paths[fileUrl || pkg.origFileUrl] = pkg;
-		pkg.fileUrl = fileUrl;
-		return source;
-	},function(ex){
-		return npmTraverseUp(context, pkg, fileUrl);
-	});
-};
-
-function npmTraverseUp(context, pkg, fileUrl) {
-	// make sure we aren't loading something we've already loaded
-	var parentAddress = parentNodeModuleAddress(fileUrl);
-	if(!parentAddress) {
-		throw new Error('Did not find ' + pkg.origFileUrl);
-	}
-	var nodeModuleAddress = parentAddress+"/"+pkg.name+"/package.json";
-	if(context.paths[nodeModuleAddress]) {
-		// already processed
-		return;
-	} else {
-		return npmLoad(context, pkg, nodeModuleAddress);
-	}
-}
 
 // Translate helpers ===============
 // Given all the package.json data, these helpers help convert it to a source.
@@ -296,18 +113,18 @@ function convertPropertyNamesAndValues (context, pkg, map , root) {
 }
 
 function convertName (context, pkg, map, root, name) {
-	var parsed = parseModuleName(name, pkg.name);
-	if(name.indexOf("#") >= 0) {
+	var parsed = utils.moduleName.parse(name, pkg.name);
+	if( name.indexOf("#") >= 0 ) {
 		
 		if(parsed.packageName === pkg.name) {
 			parsed.version = pkg.version;
 		} else {
 			// Get the requested version's actual version.
-			var requestedVersion = getDependencyMap(context.loader, pkg)[parsed.packageName].version;
+			var requestedVersion = crawl.getDependencyMap(context.loader, pkg, root)[parsed.packageName].version;
 			var depPkg = context.versions[parsed.packageName][requestedVersion];
 			parsed.version = depPkg.version;
 		}
-		return createModuleName(parsed);
+		return utils.moduleName.create(parsed);
 		
 	} else {
 		
@@ -316,7 +133,7 @@ function convertName (context, pkg, map, root, name) {
 		} else {
 			// this is for a module within the package
 			if (property.substr(0,2) === "./" ) {
-				return createModuleName({
+				return utils.moduleName.create({
 					packageName: pkg.name,
 					modulePath: name,
 					version: pkg.version,
@@ -329,7 +146,7 @@ function convertName (context, pkg, map, root, name) {
 				if(  pkg.name === parsed.packageName || ( (pkg.system && pkg.system.name) === parsed.packageName) ) {
 					depPkg = pkg;
 				} else {
-					var requestedProject = getDependencyMap(context.loader, pkg, root)[parsed.packageName];
+					var requestedProject = crawl.getDependencyMap(context.loader, pkg, root)[parsed.packageName];
 					if(!requestedProject) {
 						console.warn("WARN: Could not find ", name , "in node_modules. Ignoring.");
 						return name;
@@ -344,9 +161,9 @@ function convertName (context, pkg, map, root, name) {
 				
 				parsed.version = depPkg.version;
 				if(!parsed.modulePath) {
-					parsed.modulePath = npmExtension.pkgMain(depPkg);
+					parsed.modulePath = utils.pkg.main(depPkg);
 				}
-				return createModuleName(parsed);
+				return utils.moduleName.create(parsed);
 			}
 			
 		}
@@ -395,10 +212,10 @@ function convertBrowser(pkg, browser) {
 function convertBrowserProperty(map, pkg, fromName, toName) {
 	var packageName = pkg.name;
 	
-	var fromParsed = parseModuleName(fromName, packageName),
-		  toParsed = toName  ? parseModuleName(toName, packageName): "@empty";
+	var fromParsed = utils.moduleName.parse(fromName, packageName),
+		  toParsed = toName  ? utils.moduleName.parse(toName, packageName): "@empty";
 	
-	map[createModuleName(fromParsed)] = createModuleName(toParsed);
+	map[utils.moduleName.create(fromParsed)] = utils.moduleName.create(toParsed);
 }
