+++ package.json
@@ -15,7 +15,7 @@
     "less": "^1.7.0"
   },
   "devDependencies": {
-    "system-npm": "bitovi/system-npm#master",
+    "system-npm": "bitovi/system-npm#cleanup",
     "system-json": "bitovi/system-json#master",
     "bower": "1.3.8",
     "grunt": "~0.4.1",
@@ -30,9 +30,7 @@
     "jquery": ">2.0"
   },
   "system": {
-    "npmIgnore": {
-      "less": true
-    }
+    "npmIgnore": ["less"]
   },
   "repository": {
     "type": "git",
